change price column data type to double

+++ database/migrations/2022_07_23_203132_create_cars_table.php
@@ -18,7 +18,7 @@ class CreateCarsTable extends Migration
             $table->string('model');
             $table->string('brand');
             $table->year('year');
-            $table->bigInteger('price');
+            $table->double('price');
             $table->string('color');
             $table->string('traction');
             $table->string('type');
